+++ app/src/main/java/robotoyard/eclabs/IExecutor.java
@@ -1,8 +1,0 @@
-package robotoyard.eclabs;
-
-/**
- * Created by Pierre on 18/03/2015.
- */
-public interface IExecutor {
-    public void execute();
-}