+++ app/src/main/java/robotoyard/pm/ia/AWorld.java
@@ -1,8 +1,0 @@
-package robotoyard.pm.ia;
-
-/**
- *
- * @author Pierre Michel
- */
-public abstract class AWorld {
-}