+++ app/src/main/java/robotoyard/pm/ia/IEndCondition.java
@@ -1,9 +1,0 @@
-package robotoyard.pm.ia;
-
-/**
- *
- * @author Pierre Michel
- */
-public interface IEndCondition {
-  public boolean checkEnd(AWorld world, AGameState state);
-}