+++ app/src/main/java/robotoyard/pm/ia/IGameMove.java
@@ -1,8 +1,0 @@
-package robotoyard.pm.ia;
-
-/**
- *
- * @author usrlocal
- */
-public interface IGameMove {
-}