+++ app/src/main/java/roboyard/eclabs/IExecutor.java
@@ -1,0 +1,8 @@
+package roboyard.eclabs;
+
+/**
+ * Created by Pierre on 18/03/2015.
+ */
+public interface IExecutor {
+    public void execute();
+}