+++ app/src/main/java/roboyard/pm/ia/AWorld.java
@@ -1,0 +1,8 @@
+package roboyard.pm.ia;
+
+/**
+ *
+ * @author Pierre Michel
+ */
+public abstract class AWorld {
+}