+++ app/src/main/java/roboyard/pm/ia/IGameMove.java
@@ -1,0 +1,8 @@
+package roboyard.pm.ia;
+
+/**
+ *
+ * @author usrlocal
+ */
+public interface IGameMove {
+}